disable timers

+++ formgps_ui_field.cpp
@@ -333,9 +333,9 @@ void FormGPS::field_new_from_KML(QString field_name, QString file_name) {
             pn.longitude = pn.lonStart;
 
             sim.latitude = pn.latStart;
-            settings->setValue(SETTINGS_gps_simLatitude, pn.latStart);
+            settings->setValue(SETTINGS_gps_simLatitude, (double)pn.latStart);
             sim.longitude = pn.lonStart;
-            settings->setValue(SETTINGS_gps_simLongitude, pn.lonStart);
+            settings->setValue(SETTINGS_gps_simLongitude, (double)pn.lonStart);
         }
     pn.SetLocalMetersPerDegree();
